Example of annotation file for raw.mine, which is the output of SpectroMine.

Annotation of example data, raw.mine, in this package. It should be prepared by users. The variables are as follows:

annotation.mine

Format

A data frame with 72 rows and 7 variables.

Details

Run : MS run ID. It should be the same as R.FileName info in raw.mine

Channel : Labeling information (TMT6_126, ..., TMT6_131). The channels should be consistent with the channel columns in raw.mine.

Condition : Condition (ex. Healthy, Cancer, Time0). If the channal doesn't have sample, please add 'Empty' under Condition.

Mixture : Mixture of samples labeled with different TMT reagents, which can be analyzed in a single mass spectrometry experiment.

TechRepMixture : Technical replicate of one mixture. One mixture may have multiple technical replicates. For example, if 'TechRepMixture' = 1, 2 are the two technical replicates of one mixture, then they should match with same 'Mixture' value.

Fraction : Fraction ID. One technical replicate of one mixture may be fractionated into multiple fractions to increase the analytical depth. Then one technical replicate of one mixture should correspond to multuple fractions. For example, if 'Fraction' = 1, 2, 3 are three fractions of the first technical replicate of one TMT mixture of biological subjects, then they should have same 'TechRepMixture' and 'Mixture' value.

BioReplicate : Unique ID for biological subject. If the channal doesn't have sample, please add 'Empty' under BioReplicate

Examples

head(annotation.mine)

#>                                                     Run TechRepMixture Fraction
#> 1 ch_19Jan2017_SM-1-1_Sp-6-2_CID-OT-MS3-Short_HpH_1.raw              1        1
#> 2 ch_19Jan2017_SM-1-1_Sp-6-2_CID-OT-MS3-Short_HpH_2.raw              1        2
#> 3 ch_19Jan2017_SM-1-1_Sp-6-2_CID-OT-MS3-Short_HpH_3.raw              1        3
#> 4 ch_19Jan2017_SM-1-1_Sp-6-2_CID-OT-MS3-Short_HpH_4.raw              1        4
#> 5 ch_19Jan2017_SM-1-1_Sp-6-2_CID-OT-MS3-Short_HpH_5.raw              1        5
#> 6 ch_19Jan2017_SM-1-1_Sp-6-2_CID-OT-MS3-Short_HpH_6.raw              1        6
#>    Channel Condition Mixture BioReplicate
#> 1 TMT6_126         3       1            1
#> 2 TMT6_126         3       1            1
#> 3 TMT6_126         3       1            1
#> 4 TMT6_126         3       1            1
#> 5 TMT6_126         3       1            1
#> 6 TMT6_126         3       1            1