Visualization for explanatory data analysis - TMT experiment

To illustrate the quantitative data and quality control of MS runs, dataProcessPlotsTMT takes the quantitative data from converter functions (PDtoMSstatsTMTFormat, MaxQtoMSstatsTMTFormat, SpectroMinetoMSstatsTMTFormat) as input and generate two types of figures in pdf files as output : (1) profile plot (specify "ProfilePlot" in option type), to identify the potential sources of variation for each protein; (2) quality control plot (specify "QCPlot" in option type), to evaluate the systematic bias between MS runs.

dataProcessPlotsTMT(
  data.peptide,
  data.summarization,
  type,
  ylimUp = FALSE,
  ylimDown = FALSE,
  x.axis.size = 10,
  y.axis.size = 10,
  text.size = 4,
  text.angle = 90,
  legend.size = 7,
  dot.size.profile = 2,
  ncol.guide = 5,
  width = 10,
  height = 10,
  which.Protein = "all",
  originalPlot = TRUE,
  summaryPlot = TRUE,
  address = ""
)

Arguments

data.peptide
name of the data with peptide level, which can be the output of converter functions(PDtoMSstatsTMTFormat, MaxQtoMSstatsTMTFormat, SpectroMinetoMSstatsTMTFormat).

data.summarization
name of the data with protein-level, which can be the output of proteinSummarization function.

type
choice of visualization. "ProfilePlot" represents profile plot of log intensities across MS runs. "QCPlot" represents box plots of log intensities across channels and MS runs.

ylimUp
upper limit for y-axis in the log scale. FALSE(Default) for Profile Plot and QC Plot uses the upper limit as rounded off maximum of log2(intensities) after normalization + 3..

ylimDown
lower limit for y-axis in the log scale. FALSE(Default) for Profile Plot and QC Plot uses 0..

x.axis.size
size of x-axis labeling for "Run" and "channel in Profile Plot and QC Plot.

y.axis.size
size of y-axis labels. Default is 10.

text.size
size of labels represented each condition at the top of Profile plot and QC plot. Default is 4.

text.angle
angle of labels represented each condition at the top of Profile plot and QC plot. Default is 0.

legend.size
size of legend above Profile plot. Default is 7.

dot.size.profile
size of dots in Profile plot. Default is 2.

ncol.guide
number of columns for legends at the top of plot. Default is 5.

width
width of the saved pdf file. Default is 10.

height
height of the saved pdf file. Default is 10.

which.Protein
Protein list to draw plots. List can be names of Proteins or order numbers of Proteins. Default is "all", which generates all plots for each protein. For QC plot, "allonly" will generate one QC plot with all proteins.

originalPlot
TRUE(default) draws original profile plots, without normalization.

summaryPlot
TRUE(default) draws profile plots with protein summarization for each channel and MS run.

address
the name of folder that will store the results. Default folder is the current working directory. The other assigned folder has to be existed under the current working directory. An output pdf file is automatically created with the default name of "ProfilePlot.pdf" or "QCplot.pdf". The command address can help to specify where to store the file as well as how to modify the beginning of the file name. If address=FALSE, plot will be not saved as pdf file but showed in window.

Value

plot or pdf

Examples

data(input.pd)
quant.msstats = proteinSummarization(input.pd,
                                      method="msstats",
                                      global_norm=TRUE,
                                      reference_norm=TRUE)

#> INFO  [2021-05-12 03:04:21] ** MSstatsTMT - proteinSummarization function
#> INFO  [2021-05-12 03:04:22] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture1_01.raw ( 1  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:22] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture1_02.raw ( 2  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:23] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture2_01.raw ( 3  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:23] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture2_02.raw ( 4  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:24] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture2_03.raw ( 5  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:24] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture3_01.raw ( 6  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:24] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture3_02.raw ( 7  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:25] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture4_01.raw ( 8  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:25] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture4_02.raw ( 9  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:25] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture5_01.raw ( 10  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:26] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture5_02.raw ( 11  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:26] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture4_03.raw ( 12  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:26] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture5_03.raw ( 13  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:27] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture1_03.raw ( 14  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-12 03:04:27] Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture3_03.raw ( 15  of  15 )
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> Error in data.table::setnames(summarized_results, c("LogIntensities",     "originalRUN"), c("Abundance", "RunChannel")): Items of 'old' not found in column names: [LogIntensities, originalRUN]. Consider skip_absent=TRUE.

## Profile plot
dataProcessPlotsTMT(data.peptide=input.pd,
                   data.summarization=quant.msstats,
                   type='ProfilePlot',
                   width = 21,
                   height = 7)

#> Error in as.vector(y): nie znaleziono obiektu 'quant.msstats'

## NottoRun: QC plot
# dataProcessPlotsTMT(data.peptide=input.pd,
                    # data.summarization=quant.msstats,
                    # type='QCPlot',
                    # width = 21,
                    # height = 7)